Changes for v14.00
Updated the review window to have buttons for many common menu options. These buttons have taken the screen space that was previously reserved for displaying reliability information. Reliability information is now available through a separate "Analyze" window. Widened the listbox and buttons widths and shortened the annotations and vlc widths. Fixed a bug where underscores in filenames were not displaying in the listbox.